requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: win-64
absl-py=1.0.0=pyhd8ed1ab_0
aiohttp=3.8.1=py37hcc03f2d_0
aiosignal=1.2.0=pyhd8ed1ab_0
alabaster=0.7.12=py37_0
appdirs=1.4.4=pyhd3eb1b0_0
argcomplete=1.12.3=pyhd3eb1b0_0
argon2-cffi=20.1.0=py37h2bbff1b_1
arrow=0.13.1=py37_0
astroid=2.6.6=py37haa95532_0
async-timeout=4.0.1=pyhd8ed1ab_0
async_generator=1.10=py37h28b3542_0
asynctest=0.13.0=py_0
atomicwrites=1.4.0=py_0
attrs=21.2.0=pyhd3eb1b0_0
autopep8=1.6.0=pyhd3eb1b0_0
babel=2.9.1=pyhd3eb1b0_0
backcall=0.2.0=pyhd3eb1b0_0
bayesian-optimization=1.2.0=pypi_0
bcrypt=3.2.0=py37he774522_0
binaryornot=0.4.4=pyhd3eb1b0_1
black=19.10b0=py_0
blas=1.0=mkl
bleach=4.0.0=pyhd3eb1b0_0
blinker=1.4=py_1
boto3=1.18.21=pyhd3eb1b0_0
botocore=1.21.41=pyhd3eb1b0_1
bottleneck=1.3.2=py37h2a96729_1
brotli=1.0.9=ha925a31_2
brotlipy=0.7.0=py37h2bbff1b_1003
ca-certificates=2022.4.26=haa95532_0
cachetools=4.2.4=pyhd8ed1ab_0
certifi=2022.6.15=py37haa95532_0
cffi=1.15.0=py37h2bbff1b_0
chardet=4.0.0=py37haa95532_1003
charset-normalizer=2.0.4=pyhd3eb1b0_0
click=8.0.3=pyhd3eb1b0_0
cloudpickle=2.0.0=pyhd3eb1b0_0
colorama=0.4.4=pyhd3eb1b0_0
cookiecutter=1.7.2=pyhd3eb1b0_0
cryptography=35.0.0=py37h71e12ea_0
cudatoolkit=11.3.1=h59b6b97_2
cycler=0.11.0=pyhd3eb1b0_0
dataclasses=0.8=pyh6d0b6a4_7
debugpy=1.5.1=py37hd77b12b_0
decorator=5.1.0=pyhd3eb1b0_0
defusedxml=0.7.1=pyhd3eb1b0_0
diff-match-patch=20200713=pyhd3eb1b0_0
docutils=0.17.1=py37haa95532_1
entrypoints=0.3=py37_0
flake8=3.9.2=pyhd3eb1b0_0
flask=1.1.2=pyhd3eb1b0_0
fonttools=4.25.0=pyhd3eb1b0_0
freetype=2.10.4=hd328e21_0
frozenlist=1.2.0=py37hcc03f2d_1
google-auth=2.3.3=pyh6c4a22f_0
google-auth-oauthlib=0.4.6=pyhd8ed1ab_0
grpcio=1.42.0=py37hc60d5dd_0
gym=0.21.0=pypi_0
h5py=2.8.0=py37hf7173ca_2
hdf5=1.8.20=hac2f561_1
icc_rt=2019.0.0=h0cc432a_1
icu=58.2=ha925a31_3
idna=3.3=pyhd3eb1b0_0
imagesize=1.3.0=pyhd3eb1b0_0
importlib-metadata=4.8.2=py37haa95532_0
importlib_metadata=4.8.2=hd3eb1b0_0
inflection=0.5.1=py37haa95532_0
intel-openmp=2021.4.0=haa95532_3556
intervaltree=3.1.0=pyhd3eb1b0_0
ipykernel=6.4.1=py37haa95532_1
ipython=7.29.0=py37hd4e2768_0
ipython_genutils=0.2.0=pyhd3eb1b0_1
ipywidgets=7.6.5=pyhd8ed1ab_0
isort=5.9.3=pyhd3eb1b0_0
itsdangerous=2.0.1=pyhd3eb1b0_0
jedi=0.18.0=py37haa95532_1
jinja2=2.11.3=pyhd3eb1b0_0
jinja2-time=0.2.0=pyhd3eb1b0_2
jmespath=0.10.0=pyhd3eb1b0_0
joblib=1.1.0=pyhd3eb1b0_0
jpeg=9d=h2bbff1b_0
jsonschema=3.2.0=pyhd3eb1b0_2
jupyter=1.0.0=pypi_0
jupyter-console=6.4.0=pypi_0
jupyter_client=6.1.12=pyhd3eb1b0_0
jupyter_core=4.9.1=py37haa95532_0
jupyterlab_pygments=0.1.2=py_0
jupyterlab_widgets=1.0.2=pyhd8ed1ab_0
kaggle=1.5.12=pypi_0
keyring=23.4.0=py37haa95532_0
kiwisolver=1.3.1=py37hd77b12b_0
lazy-object-proxy=1.6.0=py37h2bbff1b_0
libiconv=1.15=h1df5818_7
libopencv=3.4.2=h20b85fd_0
libpng=1.6.37=h2a8f88b_0
libprotobuf=3.17.2=h23ce68f_1
libspatialindex=1.9.3=h6c2663c_0
libtiff=4.2.0=hd0e1b90_0
libuv=1.40.0=he774522_0
libwebp=1.2.0=h2bbff1b_0
libxml2=2.9.12=h0ad7f3c_0
libxslt=1.1.34=he774522_0
lxml=4.6.3=py37h9b66d53_0
lz4-c=1.9.3=h2bbff1b_1
m2w64-gcc-libgfortran=5.3.0=6
m2w64-gcc-libs=5.3.0=7
m2w64-gcc-libs-core=5.3.0=7
m2w64-gmp=6.1.0=2
m2w64-libwinpthread-git=5.0.0.4634.697f757=2
markdown=3.3.6=pyhd8ed1ab_0
markupsafe=1.1.1=py37hfa6e2cd_1
matplotlib=3.4.3=py37haa95532_0
matplotlib-base=3.4.3=py37h49ac443_0
matplotlib-inline=0.1.2=pyhd3eb1b0_2
mccabe=0.6.1=py37_1
mistune=0.8.4=py37hfa6e2cd_1001
mkl=2021.4.0=haa95532_640
mkl-service=2.4.0=py37h2bbff1b_0
mkl_fft=1.3.1=py37h277e83a_0
mkl_random=1.2.2=py37hf11a4ad_0
msys2-conda-epoch=20160418=1
multidict=5.2.0=py37hcc03f2d_1
munkres=1.1.4=py_0
mypy_extensions=0.4.3=py37_0
nbclient=0.5.3=pyhd3eb1b0_0
nbconvert=6.1.0=py37haa95532_0
nbformat=5.1.3=pyhd3eb1b0_0
nest-asyncio=1.5.1=pyhd3eb1b0_0
notebook=6.4.6=py37haa95532_0
numexpr=2.7.3=py37hb80d3ca_1
numpy=1.21.2=py37hfca59bb_0
numpy-base=1.21.2=py37h0829f74_0
numpydoc=1.1.0=pyhd3eb1b0_1
oauthlib=3.1.1=pyhd8ed1ab_0
olefile=0.46=py37_0
opencv=3.4.2=py37h40b0b35_0
openssl=1.1.1q=h2bbff1b_0
packaging=21.3=pyhd3eb1b0_0
pandas=1.3.4=py37h6214cd6_0
pandas-datareader=0.10.0=pyhd3eb1b0_0
pandocfilters=1.4.3=py37haa95532_1
paramiko=2.7.2=py_0
parso=0.8.2=pyhd3eb1b0_0
pathspec=0.7.0=py_0
pexpect=4.8.0=pyhd3eb1b0_3
pickleshare=0.7.5=pyhd3eb1b0_1003
pillow=8.4.0=py37hd45dc43_0
pip=21.2.4=py37haa95532_0
pluggy=1.0.0=py37haa95532_0
poyo=0.5.0=pyhd3eb1b0_0
prometheus_client=0.12.0=pyhd3eb1b0_0
prompt-toolkit=3.0.20=pyhd3eb1b0_0
protobuf=3.17.2=py37hd77b12b_0
psutil=5.8.0=py37h2bbff1b_1
ptyprocess=0.7.0=pyhd3eb1b0_2
py-opencv=3.4.2=py37hc319ecb_0
pyasn1=0.4.8=py_0
pyasn1-modules=0.2.7=py_0
pycodestyle=2.7.0=pyhd3eb1b0_0
pycparser=2.21=pyhd3eb1b0_0
pydocstyle=6.1.1=pyhd3eb1b0_0
pyflakes=2.3.1=pyhd3eb1b0_0
pygments=2.10.0=pyhd3eb1b0_0
pyjwt=2.3.0=pyhd8ed1ab_1
pylint=2.9.6=py37haa95532_1
pyls-spyder=0.4.0=pyhd3eb1b0_0
pynacl=1.4.0=py37h62dcd97_1
pyopenssl=21.0.0=pyhd3eb1b0_1
pyparsing=3.0.4=pyhd3eb1b0_0
pyqt=5.9.2=py37h6538335_2
pyrsistent=0.18.0=py37h196d8e1_0
pysocks=1.7.1=py37_1
python=3.7.11=h6244533_0
python-dateutil=2.8.2=pyhd3eb1b0_0
python-lsp-black=1.0.0=pyhd3eb1b0_0
python-lsp-jsonrpc=1.0.0=pyhd3eb1b0_0
python-lsp-server=1.2.4=pyhd3eb1b0_0
python-slugify=5.0.2=pyhd3eb1b0_0
python_abi=3.7=2_cp37m
pytorch=1.10.0=py3.7_cuda11.3_cudnn8_0
pytorch-mutex=1.0=cuda
pytz=2021.3=pyhd3eb1b0_0
pyu2f=0.1.5=pyhd8ed1ab_0
pywin32=228=py37hbaba5e8_1
pywin32-ctypes=0.2.0=py37_1001
pywinpty=0.5.7=py37_0
pyyaml=5.3.1=py37he774522_1
pyzmq=22.3.0=py37hd77b12b_2
qdarkstyle=3.0.2=pyhd3eb1b0_0
qstylizer=0.1.10=pyhd3eb1b0_0
qt=5.9.7=vc14h73c81de_0
qtawesome=1.0.3=pyhd3eb1b0_0
qtconsole=5.1.1=pyhd3eb1b0_0
qtpy=1.10.0=pyhd3eb1b0_0
regex=2021.8.3=py37h2bbff1b_0
requests=2.26.0=pyhd3eb1b0_0
requests-oauthlib=1.3.0=pyh9f0ad1d_0
rope=0.21.1=pyhd3eb1b0_0
rsa=4.8=pyhd8ed1ab_0
rtree=0.9.7=py37h2eaa2aa_1
s3transfer=0.5.0=pyhd3eb1b0_0
scikit-learn=1.0.1=py37hf11a4ad_0
scipy=1.7.1=py37hbe87c03_2
send2trash=1.8.0=pyhd3eb1b0_1
setuptools=58.0.4=py37haa95532_0
sip=4.19.8=py37h6538335_0
six=1.16.0=pyhd3eb1b0_0
snowballstemmer=2.2.0=pyhd3eb1b0_0
sortedcontainers=2.4.0=pyhd3eb1b0_0
sphinx=4.2.0=pyhd3eb1b0_1
sphinxcontrib-applehelp=1.0.2=pyhd3eb1b0_0
sphinxcontrib-devhelp=1.0.2=pyhd3eb1b0_0
sphinxcontrib-htmlhelp=2.0.0=pyhd3eb1b0_0
sphinxcontrib-jsmath=1.0.1=pyhd3eb1b0_0
sphinxcontrib-qthelp=1.0.3=pyhd3eb1b0_0
sphinxcontrib-serializinghtml=1.1.5=pyhd3eb1b0_0
spyder=5.1.5=py37haa95532_1
spyder-kernels=2.1.3=py37haa95532_0
sqlite=3.36.0=h2bbff1b_0
tabulate=0.8.10=py37haa95532_0
tensorboard=2.7.0=pyhd8ed1ab_0
tensorboard-data-server=0.6.0=py37h03978a9_1
tensorboard-plugin-wit=1.8.0=pyh44b312d_0
terminado=0.9.4=py37haa95532_0
testpath=0.5.0=pyhd3eb1b0_0
text-unidecode=1.3=pyhd3eb1b0_0
textdistance=4.2.1=pyhd3eb1b0_0
threadpoolctl=2.2.0=pyh0d69192_0
three-merge=0.1.1=pyhd3eb1b0_0
tinycss=0.4=pyhd3eb1b0_1002
tk=8.6.11=h2bbff1b_0
toml=0.10.2=pyhd3eb1b0_0
torchinfo=1.7.0=pyhd8ed1ab_0
torchvision=0.11.1=py37_cu113
tornado=6.1=py37h2bbff1b_0
tqdm=4.62.3=pyhd3eb1b0_1
traitlets=5.1.1=pyhd3eb1b0_0
typed-ast=1.4.3=py37h2bbff1b_1
typing-extensions=3.10.0.2=hd3eb1b0_0
typing_extensions=3.10.0.2=pyh06a4308_0
ujson=4.0.2=py37hd77b12b_0
unidecode=1.2.0=pyhd3eb1b0_0
urllib3=1.26.7=pyhd3eb1b0_0
vc=14.2=h21ff451_1
vs2015_runtime=14.27.29016=h5e58377_2
watchdog=2.1.6=py37haa95532_0
wcwidth=0.2.5=pyhd3eb1b0_0
webencodings=0.5.1=py37_1
werkzeug=2.0.2=pyhd3eb1b0_0
wheel=0.37.0=pyhd3eb1b0_1
whichcraft=0.6.1=pyhd3eb1b0_0
widgetsnbextension=3.5.2=py37h03978a9_1
win_inet_pton=1.1.0=py37haa95532_0
wincertstore=0.2=py37haa95532_2
winpty=0.4.3=4
wrapt=1.12.1=py37he774522_1
xz=5.2.5=h62dcd97_0
yaml=0.2.5=he774522_0
yapf=0.31.0=pyhd3eb1b0_0
yarl=1.7.2=py37hcc03f2d_1
zipp=3.6.0=pyhd3eb1b0_0
zlib=1.2.11=h62dcd97_4
zstd=1.4.9=h19a0ad4_0
